Refining GUI
This takes the old existing GUI that was imported with the JavaFX sample project and makes it sutable for the project. This adds text fiels for user input for Food name and type along with the lables for each text field notateing witch field should have the text enterd into. Also adds a submit button, along with adding a list view so that the user can also see what itmes have been added to the 'pantry'.