Can Now Add Items
This commit connects the codebehind and the main food class togeather so that not only can food objects be created, also allows for the user to set the begging quantity of the item. also allows for the user to see what items are in the pantry by using a listView to display in text their items currently in their pantry.