fixing problem with gui
for some reason in the last commit the GUI did not have most of hte options that were in the code behind. The combo box that I am using did not appear in the most recent version of what the GUI was supposed to look like. This commit aims to hopefully fix this issue and allow for the code in the codebehind class to be linked with some part of the GUI in the FXML file.