All of Part 2 is Done
So I accedently did all of part two (adding the buttons and textfields to the GUI, adding the methods to make the functionality work, and connecting the two in the codebehind class working with the model class. Any methods that were created and had highlevel logic were tested. I know I should have done this step by step but I got into the 'flow'.