Project 1 Is Finished
The Pantry can now have the quantities with in it counted and displayed to the users in the form of an information alert. there are not any checkstyle violations, and the code works as specified in the instructions, all test are worte for methods that need it, and proper feature branching development has been followd to the best of my abilities.